ppt_word_integration/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
python-pptx==0.6.21
python-docx==0.8.11
requests==2.26.0
gunicorn==20.1.0
pytest==6.2.5